Revert "Szorgalmi-02-Szakonyi-Adam"
