--- src/model/badge.model.tsx
@@ -1,8 +1,0 @@
-export interface BadgeModel
-{
-    id:number;
-    name: string;
-    description: string;
-    image: string;
-    createdAt: Date;
-}
