--- src/model/user.model.tsx
@@ -1,12 +1,0 @@
-export interface UserBadgeModel
-{
-    id:number;
-}
-export interface UserModel
-{
-    name: string;
-    id:number;
-    createdAt: Date;
-    image: string;
-    badges: UserBadgeModel[];
-}
