+++ src/models/user.model.ts
@@ -1,0 +1,11 @@
+export interface UserBadgeModel {
+  id: number;
+}
+
+export interface UserModel {
+  name: string;
+  id: number;
+  createdAt: Date;
+  image: string;
+  badges: UserBadgeModel[];
+}
