+++ src/services/user.service.ts
@@ -1,0 +1,10 @@
+import { UserModel } from "../models/user.model";
+import request, { Methods } from "../util/request";
+
+class UsersService {
+  async getUsers() {
+    return request<UserModel[]>({ method: Methods.GET, resource: "users" });
+  }
+}
+
+export const usersService = new UsersService();
