--- src/services/user.service.tsx
@@ -1,10 +1,0 @@
-import { UserModel } from "../model/user.model";
-import request, { Methods } from "../util/request";
-
-class UsersService {
-    async getUsers() {
-        return request<UserModel[]>({ method: Methods.GET, resource: "users" });
-    }
-}
-
-export const usersService = new UsersService();
